backend/requirements.txt
scikit-learn
django-cors-headers

mysqlclient
mysqlclient>=2.2

python-dotenv>=1.0
djangorestframework
pdfplumber
python-docx
spacy
en_core_web_sm


